[Refactoring] Move some more code from RefactoringActions.h to a .cpp file
